+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 - [#758](https://github.com/airblade/paper_trail/pull/758) -
   `PaperTrail.config.track_associations` getter method removed,
   use `track_associations?` instead.
+- [#740](https://github.com/airblade/paper_trail/issues/740) -
+  `PaperTrail.config.track_associations?` now defaults to false
 - [#723](https://github.com/airblade/paper_trail/pull/723) -
   `PaperTrail.enabled=` now affects all threads
 - [#556](https://github.com/airblade/paper_trail/pull/556) /